pyproject.toml
[project]
name = "lab1"
version = "0.1.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.11.0",
    "click>=8.3.1",
    "fastapi>=0.121.2",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "mlflow>=3.6.0",
    "numpy<2.0.0",
    "onnx>=1.19.0",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.0",
    "pillow>=12.0.0",
    "pylint>=4.0.3",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "python-multipart>=0.0.20",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "uvicorn>=0.38.0",
]
